Split the claim submission form per journey
Different journeys write different attributes to the eligibilities.
Splitting the forms feels a bit awkward (having to slice the appropriate
attributes), which makes me wonder if sharing the journey session
between journeys is the right approach. Maybe we're better off with a
separate session per journey.
